service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
torch
timm
transformers
